Fix every -Wall -Wextra warning across all build presets (#28)
Debug/Release editor/game builds, plus CI's Debug/Release reproduction,
now compile with zero warnings (previously ~2100, most of it duplicate
-Wunknown-pragmas noise from MSVC-only #pragma warning blocks with no
_MSC_VER guard, and vendored code sharing our -Wall -Wextra flags
because it wasn't actually treated as SYSTEM despite the CMakeLists.txt
comment claiming it was).

Most fixes are mechanical (constructor init-list order, unused
variables, sign-compare), but several turned out to be real bugs:

- VoxModel::Save: memset(pallette, 5, 256) on a 1024-byte RGBA[256]
  array left 3/4 of the palette uninitialized.
- Camera::SetOrthographic compared its parameter to itself
  (bIsOrthographic == bIsOrthographic), so the setter never applied.
- ImguiSystem::Deinitialize deleted an ImPlatform* through a
  non-virtual base destructor (UB, derived cleanup skipped).
- VKPassBindings.cpp incremented and read the same counter within one
  function call's arguments, an order-dependent sampler-label bug.
- DataHook.h defined five no-op macros unconditionally, then
  immediately shadowed them with the real ones every time - dead code
  masquerading as a feature toggle.
- Two hand-rolled position-to-hash packers (IntegrityChecker.cpp,
  PhysicsSystem.cpp) type-punned a uint64_t* as uint16_t*, which only
  surfaced as a real -Wstrict-aliasing hazard under -O2 (Release-only)
  - rewritten as portable bit-shifts.
- Several -Wunused-but-set-variable cases were locals read only inside
  assert(...), invisible in Debug but flagged once Release's NDEBUG
  strips the assert - marked [[maybe_unused]] rather than deleted, to
  keep the Debug-time guard.
- Six HLSL shaders had deprecated empty {} effect-syntax blocks and
  three had implicit float4->float3 truncations (missing .xyz
  swizzles); reverified with spirv-val after the fix.

Vendored code (STB, imgui, imguizmo, optick) is now actually silent
via -w on vendored sources and #pragma GCC system_header on vendored
headers, without touching vendored content, matching what the build's
own comments already claimed was happening.

+++ Game/Source/UI/LevelSelect/LevelSelectCanvas.cpp
@@ -167,7 +167,7 @@ void LevelSelectCanvas::SetFocusRight()
 	Canvas::SetFocusRight();
 
 	m_iCurrentlySelectedWorld++;
-	if (m_bLoopWorlds && m_iCurrentlySelectedWorld >= m_LevelSelectWorlds.size())
+	if (m_bLoopWorlds && m_iCurrentlySelectedWorld >= static_cast<int>(m_LevelSelectWorlds.size()))
 		m_iCurrentlySelectedWorld = 0;
 
 	UpdateSelectedWorld();
